Bloc3 - Fraud detection by ML/Projet/Architecture/localModel/Dockers/flask/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.3
numpy==1.26.2
mlflow==2.9.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
scikit-learn==1.3.2
